// src/constant/error-message.enum.ts
export enum ErrorMessageEnum {
  // ================= General error message ===============================
  NOT_FOUND = 'Dữ liệu không tìm thấy',
  INTERNAL_SERVER_ERROR = 'Lỗi máy chủ',
  UNAUTHORIZED = 'Không được ủy quyền',
  FORBIDDEN = 'Quyền truy cập bị từ chối',
  BAD_REQUEST = 'Yêu cầu không hợp lệ',
  SUCCESS = 'Thành công',
  NOT_ACCEPTABLE = 'Không có quyền truy cập',
  INVALID_STATUS = 'Trạng thái không hợp lệ',
  NAME_ALREADY_EXISTS = 'Dữ liệu Tên đã tồn tại',
  INVALID_QUANTITY = 'Số lượng không hợp lệ',
  UPDATE_FAILED = 'Sửa không thành công',
  CAN_NOT_DELETE = 'Bản ghi này không thể xóa',
  CAN_NOT_UPDATE = 'Bản ghi này không thể sửa',
  CODE_EXIST = 'Dữ liệu Mã đã tồn tại',
}